pyproject.toml
[project]
name = "mcp-toolbox"
version = "0.2.0"
description = "Generate production-ready MCP servers from OpenAPI specs in one command."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
keywords = ["mcp", "openapi", "code-generation", "ai-agent", "model-context-protocol"]
classifiers = [
 "Development Status :: 3 - Alpha",
 "Intended Audience :: Developers",
 "License :: OSI Approved :: MIT License",
 "Programming Language :: Python :: 3",
 "Programming Language :: Python :: 3.10",
 "Programming Language :: Python :: 3.11",
 "Programming Language :: Python :: 3.12",
 "Programming Language :: Python :: 3.13",
 "Topic :: Software Development :: Code Generators",
]
dependencies = [
 "typer>=0.12",
 "rich>=13.0",
 "jinja2>=3.1",
 "pyyaml>=6.0",
 "openapi-spec-validator>=0.8",
 "prance>=23.6.21.0",
 "openapi-pydantic>=0.5",
 "pydantic>=2.5",
 "pydantic-settings>=2.1",
 "jsonref>=1.1",
 "chardet>=5.2,<6",
]

[project.optional-dependencies]
dev = [
 "pytest>=8.0",
 "pytest-asyncio>=0.23",
 "ruff>=0.4",
 "mcp[cli]>=1.26.0",
 "respx>=0.21",
 "tomli>=2.0; python_version < '3.11'",
]

[project.scripts]
mcp-toolbox = "mcp_toolbox.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mcp_toolbox"]

[tool.mcp_toolbox]
default_max_tools = 12
default_transport = "stdio"
default_body_style = "auto"
default_enum_style = "literal"
watch_interval_seconds = 1.0
